requirements.txt
streamlit
groq
python-dotenv
pandas
matplotlib
fsspec
huggingface_hub